Move lots of PFE interfaces to Types.ts
