+++ dev/src/codewind/Types.ts
@@ -9,17 +9,112 @@
  *     IBM Corporation - initial API and implementation
  *******************************************************************************/
 
-///// Project
+///// This file declares types for request/response bodies when communicating with the Codewind server, or with cwctl.
 
-export interface MetricsDashboardStatus {
-    hosting: "project" | "performanceContainer" | null;
-    path: string | null;
+/**
+ * Response from 'cwctl sectoken get'. Use the access_token to get past the Gatekeeper.
+ */
+export interface AccessToken {
+    readonly access_token: string;
+    readonly expires_in: number;
+    readonly token_type: string;
 }
 
-export interface MetricsInjectionStatus {
-    injectable: boolean;
-    injected: boolean;
+/**
+ * Status of Local Codewind
+ */
+export interface CLIStatus {
+    // status: "uninstalled" | "stopped" | "started";
+    "installed-versions": string[];
+    started: string[];
+    url?: string;   // only set when started
 }
+
+///// PFE Environment /////
+
+// From https://github.com/eclipse/codewind/blob/master/src/pfe/portal/routes/environment.route.js
+export interface RawCWEnvData {
+    readonly codewind_version?: string;
+    readonly image_build_time?: string;
+    readonly namespace?: string;
+    readonly os_platform: string;
+    readonly running_in_k8s: boolean;
+    readonly socket_namespace?: string;
+    readonly tekton_dashboard: TektonStatus;
+}
+
+/**
+ * Massaged env data, which the plugin is actually interested in
+ */
+export interface CWEnvData {
+    // readonly workspace: string;
+    readonly namespace?: string;
+    readonly socketNamespace: string;
+    readonly version: string;
+    readonly tektonStatus: TektonStatus;
+}
+
+interface TektonStatus {
+    readonly status: boolean;
+    readonly message: string;       // error message or "not-installed" if status is false
+    readonly url: string;           // empty if status is false
+}
+
+// From https://github.com/eclipse/codewind/blob/master/src/pfe/portal/modules/utils/Logger.js#L38
+export interface PFELogLevels {
+    readonly currentLevel: string;
+    readonly defaultLevel: string;
+    readonly allLevels: string[];
+}
+
+///// Connection /////
+
+export interface CLIConnectionData {
+    readonly id: string;
+    readonly label: string;
+    readonly url: string;
+    readonly username: string;
+    // These 3 are provided in the 'connections list' output but are not yet consumed.
+    readonly auth?: string;
+    readonly realm?: string;
+    readonly clientid?: string;
+}
+
+/**
+ * Template repository/source data as provided by the backend
+ */
+export interface TemplateSource {
+    readonly url: string;
+    readonly name?: string;
+    readonly description?: string;
+    readonly enabled: boolean;
+    readonly projectStyles: string[];
+    readonly protected: boolean;
+}
+
+/**
+ * 'data' field of ENABLE_DISABLE event, which can be converted to an enablement request.
+ */
+export interface SourceEnablement {
+    readonly repos: Array<{
+        readonly repoID: string;
+        readonly enable: boolean;
+    }>;
+}
+
+/**
+ * Represents a template/stack.
+ */
+export interface CWTemplateData {
+    label: string;
+    description: string;
+    url: string;
+    language: string;
+    projectType: string;
+    source?: string;
+}
+
+///// Project /////
 
 /**
  * Project data as
@@ -70,4 +165,39 @@ export interface PFEProjectData {
     readonly metricsDashboard?: MetricsDashboardStatus;
     readonly perfDashboardPath?: string | null;
     readonly injection?: MetricsInjectionStatus;
+}
+
+
+export interface MetricsDashboardStatus {
+    hosting: "project" | "performanceContainer" | null;
+    path: string | null;
+}
+
+export interface MetricsInjectionStatus {
+    injectable: boolean;
+    injected: boolean;
+}
+
+export interface ILogResponse {
+    readonly build?: ILogObject[];
+    readonly app?: ILogObject[];
+}
+
+export interface ILogObject {
+    readonly logName: string;
+    readonly workspaceLogPath?: string;
+}
+
+///// Project creation /////
+
+export interface IDetectedProjectType {
+    language: string;
+    projectType: string;
+    projectSubtype?: string;
+}
+
+export interface IInitializationResponse {
+    status: string;
+    result: IDetectedProjectType | string | { error: string };
+    projectPath?: string;
 }
