+++ dev/src/codewind/connection/ConnectionMemento.ts
@@ -14,8 +14,9 @@ import * as vscode from "vscode";
 import Log from "../../Logger";
 import ConnectionManager from "./ConnectionManager";
 import MCUtil from "../../MCUtil";
-import { CLICommandRunner, CLIConnectionData } from "./CLICommandRunner";
+import { CLICommandRunner } from "./CLICommandRunner";
 import remoteConnectionOverviewCmd from "../../command/connection/ConnectionOverviewCmd";
+import { CLIConnectionData } from "../Types";
 
 export namespace ConnectionMemento {
